requirements.txt
pyyaml
numpy
scipy
torch
torchvision
pillow
rouge==1.0.1
transformers==4.38.0
tiktoken
openai
anthropic
google-generativeai
opencv-python
huggingface-hub>=0.34.0,<1.0